extracts advanced search to controller concern
