+++ app/controllers/proposals_controller.rb
@@ -1,11 +1,9 @@
 class ProposalsController < ApplicationController
   include CommentableActions
   include FlagActions
+  include Search
 
-  before_action :parse_search_terms, only: [:index, :suggest]
-  before_action :parse_advanced_search_terms, only: :index
   before_action :parse_tag_filter, only: :index
-  before_action :set_search_order, only: :index
   before_action :load_categories, only: [:index, :new, :edit, :map, :summary]
   before_action :load_geozones, only: [:edit, :map, :summary]
   before_action :authenticate_user!, except: [:index, :show, :map, :summary]
